# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Redstone Squid API",
    "description": "Versioned public API for the Redstone Squid build catalog.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/verify": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Get Verification Code",
        "description": "Generate a verification code for a user.",
        "operationId": "get_verification_code_v1_verify_post",
        "security": [
          {
            "ApiCredential": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255,
                  "pattern": "^[\\x21-\\x7e]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Deduplicate an equivalent mutation by this authenticated caller for 24 hours.",
              "title": "Idempotency-Key"
            },
            "description": "Deduplicate an equivalent mutation by this authenticated caller for 24 hours."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Get Verification Code V1 Verify Post"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/verify": {
      "post": {
        "summary": "Get Verification Code",
        "description": "Generate a verification code for a user.",
        "operationId": "get_verification_code_verify_post",
        "security": [
          {
            "ApiCredential": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255,
                  "pattern": "^[\\x21-\\x7e]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Deduplicate an equivalent mutation by this authenticated caller for 24 hours.",
              "title": "Idempotency-Key"
            },
            "description": "Deduplicate an equivalent mutation by this authenticated caller for 24 hours."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Get Verification Code Verify Post"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "ACCOUNT_ALREADY_LINKED",
          "ALIAS_ALREADY_ADDED",
          "ALIAS_ALREADY_CLAIMED",
          "ALIAS_IN_USE",
          "BUILD_BUSY",
          "BUILD_NOT_FOUND",
          "BUILD_REVISION_MISMATCH",
          "BUILD_REVISION_REQUIRED",
          "CLAIM_NOT_FOUND",
          "CREATOR_ALIAS_NOT_FOUND",
          "CONFIGURATION_ERROR",
          "CONSENT_REQUIRED",
          "DATA_INTEGRITY_ERROR",
          "DOMAIN_ERROR",
          "INFRASTRUCTURE_ERROR",
          "IDEMPOTENCY_CONFLICT",
          "IDEMPOTENCY_IN_PROGRESS",
          "INTERNAL_ERROR",
          "INVALID_BUILD",
          "INVALID_CURSOR",
          "INVALID_MESSAGE",
          "INVALID_QUERY",
          "INVALID_REQUEST",
          "INVALID_STATE",
          "INVALID_USER",
          "INVALID_VERIFICATION_CODE",
          "INVALID_VERSION",
          "INVALID_VOTE_CONFIGURATION",
          "MESSAGE_NOT_FOUND",
          "MINECRAFT_ACCOUNT_NOT_FOUND",
          "MINECRAFT_SERVICE_UNAVAILABLE",
          "NOT_FOUND",
          "PERSISTENCE_ERROR",
          "RATE_LIMITED",
          "RESTRICTION_NOT_FOUND",
          "SCHEMATIC_INVALID",
          "SCHEMATIC_NOT_FOUND",
          "SCHEMATIC_RENDER_UNAVAILABLE",
          "SCHEMATIC_SUPPORT_UNAVAILABLE",
          "SCHEMATIC_TIMEOUT",
          "SCHEMATIC_TOO_LARGE",
          "SCHEMATIC_WORKER_CRASHED",
          "UNAUTHORIZED",
          "USER_NOT_FOUND",
          "VALIDATION_ERROR",
          "VERSION_CATALOG_UNAVAILABLE"
        ],
        "title": "ErrorCode",
        "description": "Stable machine-readable application error codes."
      },
...
      "JSONValue": {
        "anyOf": [
          {
            "type": "boolean"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "string"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JSONValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "resource": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JSONValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "error_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "description": "RFC 9457 problem detail response with application extensions."
      },
...
      "User": {
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Uuid"
          }
        },
        "type": "object",
        "required": [
          "uuid"
        ],
        "title": "User",
        "description": "A user model."
      },
...
    },
    "securitySchemes": {
      "ApiCredential": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
...
}